[refactor] 약속 만들기 페이지 리팩토링
